requirements.txt
streamlit==1.25.0
yt-dlp==2024.4.4
ffmpeg-python==0.2.0
whisper==1.0
transformers==4.32.0
torch==2.0.1
